refactor(core): add getCallControlsPhase and shared SpaceGroupCallState
CodeRabbit: centralize call-state buckets for in-call UI; move the
SpaceGroupCallState union to a cycle-free module for reuse.
