+++ packages/core/src/matrix/client/hooks/index.ts
@@ -2,6 +2,7 @@ export * from './use-matrix-token';
 export * from './use-user-privy-id-by-matrix-id';
 export * from './use-matrix-user-ids-by-privy-subs';
 export * from './use-space-group-call';
+export { getCallControlsPhase } from './space-group-call-state';
 export { isPermissionLikeGroupCallError } from './space-group-call-utils';
 export { logSpaceGroupCallEvent } from './space-group-call-telemetry';
 export type { SpaceGroupCallTelemetryEvent } from './space-group-call-telemetry';
